+++ CHANGELOG.rst
@@ -2,8 +2,28 @@
 Change Log
 ==========
 
+v0.0.4
+------
+
+* improvements for error handling; added click for command line
+
+
+v0.0.3
+------
+
+* improvements to NetworkManager script
+
+
+v0.0.2
+------
+
+* adapted form inputs to new form layout with checkbox
+
+
 v0.0.1
 ------
 
 * initial release
+
+